+++ app/exceptions/boat_document_errors.ts
@@ -1,0 +1,6 @@
+export class BoatDocumentNotFoundError extends Error {
+  constructor(id: number) {
+    super(`BoatDocument not found: ${id}`)
+    this.name = 'BoatDocumentNotFoundError'
+  }
+}